+++ tse_analytics/modules/phenomaster/actimot/views/actimot_heatmap_plot_widget.py
@@ -26,7 +26,7 @@ class ActimotHeatmapPlotWidget(QWidget):
 
     def set_data(self, trj_df: traja.TrajaDataFrame) -> None:
         self.trj_df = trj_df
-        # self._update_plot()
+        # self.update_plot()
 
     def _update_plot(self):
         if self.trj_df is None: