+++ tse_analytics/views/data/data_plot_widget.py
@@ -52,15 +52,10 @@ class DataPlotWidget(QWidget, MessengerListener):
         self.ui.widgetSettings.layout().addWidget(self.plot_toolbar)
         self.plot_toolbar.hide()
 
-        self.ui.groupBoxDisplayErrors.hide()
-
     def register_to_messenger(self, messenger: Messenger):
         messenger.subscribe(self, DatasetChangedMessage, self._on_dataset_changed)
         messenger.subscribe(self, BinningMessage, self._on_binning_applied)
         messenger.subscribe(self, DataChangedMessage, self._on_data_changed)
-
-    def _variable_changed(self, variable: str) -> None:
-        self._assign_data()
 
     def _get_split_mode(self) -> SplitMode:
         if self.ui.radioButtonSplitByAnimal.isChecked():
@@ -72,36 +67,27 @@ class DataPlotWidget(QWidget, MessengerListener):
         else:
             return SplitMode.TOTAL
 
+    def _variable_changed(self, variable: str) -> None:
+        self._refresh_data()
+
     def _factor_changed(self, factor_name: str) -> None:
-        split_mode = self._get_split_mode()
-        selected_factor_name = self.ui.factorSelector.currentText()
-        factor = Manager.data.selected_dataset.factors[selected_factor_name] if selected_factor_name != "" else None
-        self.timelinePlotView.set_split_mode(split_mode, factor)
-        self.barPlotView.set_split_mode(split_mode, factor)
-        self._assign_data()
+        self._refresh_data()
 
     def _error_type_std_toggled(self, toggled: bool) -> None:
         if not toggled:
             return
-        if Manager.data.selected_dataset.binning_settings.mode == BinningMode.INTERVALS:
-            self.timelinePlotView.set_error_type("std")
-        else:
-            self.barPlotView.set_error_type("sd")
+        self._refresh_data()
 
     def _error_type_ste_toggled(self, toggled: bool) -> None:
         if not toggled:
             return
-        if Manager.data.selected_dataset.binning_settings.mode == BinningMode.INTERVALS:
-            self.timelinePlotView.set_error_type("sem")
-        else:
-            self.barPlotView.set_error_type("se")
+        self._refresh_data()
 
     def _display_errors_toggled(self, toggled: bool) -> None:
-        self.timelinePlotView.set_display_errors(toggled)
-        self.barPlotView.set_display_errors(toggled)
+        self._refresh_data()
 
     def _set_scatter_plot(self, state: bool):
-        self.timelinePlotView.set_scatter_plot(state)
+        self._refresh_data()
 
     def _on_dataset_changed(self, message: DatasetChangedMessage):
         self.ui.pushButtonAddReport.setDisabled(message.dataset is None)
@@ -113,29 +99,21 @@ class DataPlotWidget(QWidget, MessengerListener):
         else:
             self.ui.variableSelector.set_data(message.dataset.variables)
             self.ui.factorSelector.set_data(message.dataset.factors, add_empty_item=False)
-            self._assign_data()
+            self._refresh_data()
 
     def _on_binning_applied(self, message: BinningMessage):
-        self._assign_data()
+        self._refresh_data()
 
     def _on_data_changed(self, message: DataChangedMessage):
-        self._assign_data()
+        self._refresh_data()
 
     def _split_mode_toggled(self, toggled: bool):
         if not toggled:
             return
+        self.ui.factorSelector.setEnabled(self.ui.radioButtonSplitByFactor.isChecked())
+        self._refresh_data()
 
-        selected_factor_name = self.ui.factorSelector.currentText()
-        split_mode = self._get_split_mode()
-
-        self.ui.factorSelector.setEnabled(split_mode == SplitMode.FACTOR)
-
-        factor = Manager.data.selected_dataset.factors[selected_factor_name] if selected_factor_name != "" else None
-        self.timelinePlotView.set_split_mode(split_mode, factor)
-        self.barPlotView.set_split_mode(split_mode, factor)
-        self._assign_data()
-
-    def _assign_data(self):
+    def _refresh_data(self):
         if Manager.data.selected_dataset is None:
             return
 
@@ -154,48 +132,88 @@ class DataPlotWidget(QWidget, MessengerListener):
             return
 
         selected_variable = self.ui.variableSelector.get_selected_variable()
+        display_errors = self.ui.groupBoxDisplayErrors.isChecked()
 
         if not Manager.data.selected_dataset.binning_settings.apply:
-            self._display_timeline_plot(selected_variable, split_mode, selected_factor_name)
+            self._display_timeline_plot(selected_variable, split_mode, selected_factor_name, display_errors)
         else:
             if Manager.data.selected_dataset.binning_settings.mode == BinningMode.INTERVALS:
-                self._display_timeline_plot(selected_variable, split_mode, selected_factor_name)
+                self._display_timeline_plot(selected_variable, split_mode, selected_factor_name, display_errors)
             else:
-                self._display_bar_plot(selected_variable)
+                self._display_bar_plot(selected_variable, split_mode, selected_factor_name, display_errors)
 
     def _display_timeline_plot(
         self,
         selected_variable: Variable,
         split_mode: SplitMode,
         selected_factor_name: str,
+        display_errors: bool,
     ):
         self.ui.splitter.replaceWidget(0, self.timelinePlotView)
         self.barPlotView.hide()
         self.timelinePlotView.show()
+        self.plot_toolbar.hide()
+        self.ui.checkBoxScatterPlot.show()
+
+        if display_errors:
+            calculate_errors = "sem" if self.ui.radioButtonStandardError.isChecked() else "std"
+        else:
+            calculate_errors = None
+
         df = Manager.data.get_timeline_plot_df(
             variable=selected_variable,
             split_mode=split_mode,
             selected_factor_name=selected_factor_name,
+            calculate_errors=calculate_errors,
         )
-        self.timelinePlotView.set_variable(selected_variable, False)
-        self.timelinePlotView.set_data(df)
-        self.plot_toolbar.hide()
-        self.ui.checkBoxScatterPlot.show()
+
+        selected_factor = (
+            Manager.data.selected_dataset.factors[selected_factor_name] if selected_factor_name != "" else None
+        )
+
+        self.timelinePlotView.refresh_data(
+            df,
+            selected_variable,
+            split_mode,
+            display_errors,
+            selected_factor,
+            self.ui.checkBoxScatterPlot.isChecked(),
+        )
 
     def _display_bar_plot(
         self,
         selected_variable: Variable,
+        split_mode: SplitMode,
+        selected_factor_name: str,
+        display_errors: bool,
     ):
         self.ui.splitter.replaceWidget(0, self.barPlotView)
         self.timelinePlotView.hide()
         self.barPlotView.show()
+        self.ui.checkBoxScatterPlot.hide()
+
+        if display_errors:
+            calculate_errors = "se" if self.ui.radioButtonStandardError.isChecked() else "sd"
+        else:
+            calculate_errors = None
+
         df = Manager.data.get_bar_plot_df(
             variable=selected_variable,
         )
-        self.barPlotView.set_variable(selected_variable, False)
-        self.barPlotView.set_data(df)
 
-        self.ui.checkBoxScatterPlot.hide()
+        selected_factor = (
+            Manager.data.selected_dataset.factors[selected_factor_name] if selected_factor_name != "" else None
+        )
+
+        self.barPlotView.refresh_data(
+            df,
+            selected_variable,
+            split_mode,
+            selected_factor,
+            display_errors,
+            calculate_errors,
+        )
+
         new_toolbar = NavigationToolbar2QT(self.barPlotView.canvas, self)
         new_toolbar.setIconSize(QSize(16, 16))
         self.plot_toolbar.hide()