+++ tse_analytics/views/data/timeline_plot_view.py
@@ -60,42 +60,28 @@ class TimelinePlotView(pg.GraphicsLayoutWidget):
         rgn = viewRange[0]
         self.region.setRegion(rgn)
 
-    def set_data(self, df: pd.DataFrame) -> None:
+    def refresh_data(
+        self,
+        df: pd.DataFrame,
+        variable: Variable,
+        split_mode: SplitMode,
+        display_errors: bool,
+        selected_factor: Factor,
+        scatter_plot: bool,
+    ) -> None:
         self._df = df
-
-        if self._df.empty:
-            self.p1.clear()
-            self.p2.clearPlots()
-            self.legend.clear()
-            return
+        self._variable = variable
+        self._split_mode = split_mode
+        self._display_errors = display_errors
+        self._selected_factor = selected_factor
+        self._scatter_plot = scatter_plot
 
         unique_deltas = self._df["Timedelta"].unique()
         TimedeltaAxisItem.sampling_interval = unique_deltas[1] - unique_deltas[0]
 
-        self._update_plot()
+        self.update_plot()
 
-    def set_variable(self, variable: Variable, update: bool):
-        self._variable = variable
-        if update:
-            self._update_plot()
-
-    def set_split_mode(self, grouping_mode: SplitMode, selected_factor: Factor):
-        self._split_mode = grouping_mode
-        self._selected_factor = selected_factor
-
-    def set_display_errors(self, state: bool):
-        self._display_errors = state
-        self._update_plot()
-
-    def set_error_type(self, error_type: str):
-        self._error_type = error_type
-        self._update_plot()
-
-    def set_scatter_plot(self, state: bool):
-        self._scatter_plot = state
-        self._update_plot()
-
-    def _update_plot(self):
+    def update_plot(self):
         self.p1.clear()
         self.p2.clearPlots()
         self.legend.clear()
@@ -123,6 +109,11 @@ class TimelinePlotView(pg.GraphicsLayoutWidget):
 
         self.update()
 
+    def set_variable(self, variable: Variable, update: bool):
+        self._variable = variable
+        if update:
+            self.update_plot()
+
     def _plot_item(self, data: pd.DataFrame, name: str, pen):
         # x = (data["DateTime"] - pd.Timestamp("1970-01-01")) // pd.Timedelta("1s")  # Convert to POSIX timestamp
         x = data["Bin"]
@@ -138,7 +129,7 @@ class TimelinePlotView(pg.GraphicsLayoutWidget):
         p1d = self.p1.scatterPlot(x, y, pen=pen, size=2) if self._scatter_plot else self.p1.plot(x, y, pen=pen)
         self.p1.setTitle(self._variable.name)
 
-        if self._display_errors and self._split_mode != SplitMode.ANIMAL:
+        if self._display_errors and "Error" in data.columns:
             # Error bars
             error_plot = pg.ErrorBarItem(beam=0.2)
             error = data["Error"].to_numpy()
@@ -178,28 +169,9 @@ class TimelinePlotView(pg.GraphicsLayoutWidget):
 
         factor_name = self._selected_factor.name
 
-        # group_by = ["Bin", factor_name]
-        # grouped = self._df.groupby(group_by, dropna=False, observed=False)
-        #
-        # result = (
-        #     grouped.agg(
-        #         Value=(self._variable.name, self._variable.aggregation),
-        #         Error=(self._variable.name, self._error_type),
-        #     )
-        #     if self._display_errors
-        #     else grouped.agg(
-        #         Value=(self._variable.name, self._variable.aggregation),
-        #     )
-        # )
-        #
-        # data = result.reset_index()
-        # data.rename(columns={"Value": self._variable.name}, inplace=True)
-
-        data = self._df
-
         groups = self._selected_factor.groups
         for i, group in enumerate(groups):
-            filtered_data = data[data[factor_name] == group.name]
+            filtered_data = self._df[self._df[factor_name] == group.name]
 
             pen = mkPen(color=(i, len(groups)), width=1)
             tmp_min, tmp_max = self._plot_item(filtered_data, f"{group.name}", pen)
@@ -215,28 +187,9 @@ class TimelinePlotView(pg.GraphicsLayoutWidget):
         x_min = None
         x_max = None
 
-        # group_by = ["Bin", "Run"]
-        # grouped = self._df.groupby(group_by, dropna=False, observed=False)
-        #
-        # result = (
-        #     grouped.agg(
-        #         Value=(self._variable.name, self._variable.aggregation),
-        #         Error=(self._variable.name, np.std),
-        #     )
-        #     if self._display_errors
-        #     else grouped.agg(
-        #         Value=(self._variable.name, self._variable.aggregation),
-        #     )
-        # )
-        #
-        # data = result.reset_index()
-        # data.rename(columns={"Value": self._variable.name}, inplace=True)
-
-        data = self._df
-
-        runs = data["Run"].unique()
+        runs = self._df["Run"].unique()
         for i, run in enumerate(runs):
-            filtered_data = data[data["Run"] == run]
+            filtered_data = self._df[self._df["Run"] == run]
 
             pen = mkPen(color=(i, len(runs)), width=1)
             tmp_min, tmp_max = self._plot_item(filtered_data, f"Run {run}", pen)
@@ -252,27 +205,8 @@ class TimelinePlotView(pg.GraphicsLayoutWidget):
         x_min = None
         x_max = None
 
-        # group_by = ["Bin"]
-        # grouped = self._df.groupby(group_by, dropna=False, observed=False)
-        #
-        # result = (
-        #     grouped.agg(
-        #         Value=(self._variable.name, self._variable.aggregation),
-        #         Error=(self._variable.name, self._error_type),
-        #     )
-        #     if self._display_errors
-        #     else grouped.agg(
-        #         Value=(self._variable.name, self._variable.aggregation),
-        #     )
-        # )
-        #
-        # data = result.reset_index()
-        # data.rename(columns={"Value": self._variable.name}, inplace=True)
-
-        data = self._df
-
         pen = mkPen(color=(1, 1), width=1)
-        tmp_min, tmp_max = self._plot_item(data, "Total", pen)
+        tmp_min, tmp_max = self._plot_item(self._df, "Total", pen)
 
         if x_min is None or tmp_min < x_min:
             x_min = tmp_min